More edits
- added language code to Language class (JAP, ENG-US, etc)
- the course number is built from the language code of the alphabet and
the course id (JAP-1, etc)
- Courses and Decks are now both related to an Alphabet. The idea
behind this is that when you create a Course or Deck you have to choose
an Alphabet so that this will limit your options to choose from when
you add decks to Chapters or Flashcards to Decks. Really when you
shouldn’t be able to create a Course that has Chapters from all
different Alphabets. One thing that I’m thinking is that maybe the
constraint should not be by Alphabet but by Language. We need to
confirm with client.

- Searching for Courses is now working as expected. You can search by
Language or by keyword (title or description). Result list is not
paginated yet.

- I installed the UML plugin which generates the UML for you based on
the domain objects. This will allow us to have a UML diagram that’s
always synchronized. However, we still keep all the relationships in
the google doc, so any design decisions start there, then go to the
code and will then appear in the automatically generated diagram. To
view it, after you started up the app, browse to:
http://localhost:8080/OpenFluency/classDomainUML